1.6.3
Since this initial project used the 1.6.2 version simply edited for offline usage via localhost and just as this repository was added the original author updated to version 1.6.3, I opted as soon as my free time allowed after such to simply do the same with the 1.6.3 version simply edited for offline usage via localhost. Also added an "Archived" dir simply to house past and present versions in an organized fashion.